requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
seleniumbase==4.44.7
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.0.8
fastapi-limiter==0.1.6
httpx==0.27.0
openpyxl==3.1.2